+++ build_web_compilers/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.0.1
+
+- Allow the compiler to choose the default for the native null assertions
+  option when one is not explicitly provided.
+
 ## 4.0.0
 
 - Require Dart 3.0.